test(oauth2): fix protocol case and split internal hash test
