+++ coderd/oauth2provider/authorize_test.go
@@ -1,8 +1,6 @@
-package oauth2provider
+package oauth2provider_test
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -13,52 +11,10 @@ import (
 	"github.com/coder/coder/v2/site"
 )
 
-func TestHashOAuth2State(t *testing.T) {
-	t.Parallel()
-
-	t.Run("EmptyState", func(t *testing.T) {
-		t.Parallel()
-		result := hashOAuth2State("")
-		assert.False(t, result.Valid, "empty state should return invalid NullString")
-		assert.Empty(t, result.String, "empty state should return empty string")
-	})
-
-	t.Run("NonEmptyState", func(t *testing.T) {
-		t.Parallel()
-		state := "test-state-value"
-		result := hashOAuth2State(state)
-		require.True(t, result.Valid, "non-empty state should return valid NullString")
-
-		// Verify it's a proper SHA-256 hash.
-		expected := sha256.Sum256([]byte(state))
-		assert.Equal(t, hex.EncodeToString(expected[:]), result.String,
-			"state hash should be SHA-256 hex digest")
-	})
-
-	t.Run("DifferentStatesProduceDifferentHashes", func(t *testing.T) {
-		t.Parallel()
-		hash1 := hashOAuth2State("state-a")
-		hash2 := hashOAuth2State("state-b")
-		require.True(t, hash1.Valid)
-		require.True(t, hash2.Valid)
-		assert.NotEqual(t, hash1.String, hash2.String,
-			"different states should produce different hashes")
-	})
-
-	t.Run("SameStateProducesSameHash", func(t *testing.T) {
-		t.Parallel()
-		hash1 := hashOAuth2State("deterministic")
-		hash2 := hashOAuth2State("deterministic")
-		require.True(t, hash1.Valid)
-		assert.Equal(t, hash1.String, hash2.String,
-			"same state should produce identical hash")
-	})
-}
-
 func TestOAuthConsentFormIncludesCSRFToken(t *testing.T) {
 	t.Parallel()
 
-	const csrfToken = "csrf-token-value"
+	const csrfFieldValue = "csrf-field-value"
 	req := httptest.NewRequest(http.MethodGet, "https://coder.com/oauth2/authorize", nil)
 	rec := httptest.NewRecorder()
 
@@ -66,11 +22,11 @@ func TestOAuthConsentFormIncludesCSRFToken(t *testing.T) {
 		AppName:     "Test OAuth App",
 		CancelURI:   "https://coder.com/cancel",
 		RedirectURI: "https://coder.com/oauth2/authorize?client_id=test",
-		CSRFToken:   csrfToken,
+		CSRFToken:   csrfFieldValue,
 		Username:    "test-user",
 	})
 
 	require.Equal(t, http.StatusOK, rec.Result().StatusCode)
 	assert.Contains(t, rec.Body.String(), `name="csrf_token"`)
-	assert.Contains(t, rec.Body.String(), `value="`+csrfToken+`"`)
+	assert.Contains(t, rec.Body.String(), `value="`+csrfFieldValue+`"`)
 }
